Fixed the forward_model only thing left is to change the paths to local instead of hard coded!
